Cleanup around refactor
